+++ police/locales/fr.lua
@@ -60,6 +60,10 @@ i18n.importData("fr", {
 	garage_loading = "~b~Chargement...",
 	
 	cloackroom_global_title = "Vestiaire Police",
+    cloackroom_take_service_ranger_title = "Clock in as Park Ranger",
+	cloackroom_take_service_sheriff_title = "Clock in as Sheriff",
+	cloackroom_take_service_chp_title = "Clock in as State Trooper",
+    cloackroom_take_service_prison_title = "Clock in as a Prison Guard",
 	cloackroom_take_service_normal_title = "Prendre le service (uniforme)",
 	cloackroom_take_service_hidden_title = "Prendre le service (BAC)",
 	cloackroom_take_service_swat_title = "Prendre le service (intervention)",